backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiofiles==23.2.1
sqlmodel==0.0.14
sqlalchemy==2.0.23
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
tencentcloud-sdk-python==3.0.1090
cos-python-sdk-v5==1.9.30
openai==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
